backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.0",
    "google-api-python-client>=2.176.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "jira>=3.8.0",
    "langchain>=0.3.26",
    "langchain-groq>=0.3.5",
    "langgraph>=0.5.1",
    "passlib[bcrypt]>=1.7.4",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-mock>=3.14.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "tiktoken>=0.9.0",
    "uvicorn>=0.35.0",
]
